% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.vdoe.reading.scores-package.R
\docType{package}
\name{dc.vdoe.reading.scores-package}
\alias{dc.vdoe.reading.scores}
\alias{dc.vdoe.reading.scores-package}

\title{dc.vdoe.reading.scores: Social Data Commons: Education & Training: Reading Score SOL Mean and Median Pass Rate}
\description{
Allows user to easily get 3rd grade SOL reading pass rate data from the SDAD Data Commons. Data can be provided directly or in a file.

}
\author{
\strong{Maintainer}: Hanna Charankevich \email{hc2cc@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
